工作/AI能力画像文档/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
pymysql>=1.1.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
celery>=5.3.0
redis>=5.0.0
requests>=2.31.0
python-multipart>=0.0.6
